Moved trees to the afterPlace method
