+++ src/main/java/com/gildedgames/aether/world/structurepiece/GoldDungeonPieces.java
@@ -126,13 +126,6 @@ public class GoldDungeonPieces {
         private static StructurePlaceSettings makeSettings() {
             return new StructurePlaceSettings().addProcessor(VerticalGradientProcessor.INSTANCE);
         }
-
-        @Override
-        public void postProcess(WorldGenLevel level, StructureManager structureManager, ChunkGenerator generator, RandomSource random, BoundingBox chunkBox, ChunkPos pChunkPos, BlockPos pPos) {
-            super.postProcess(level, structureManager, generator, random, chunkBox, pChunkPos, pPos);
-            placeGoldenOaks(level, generator, random, this.getBoundingBox(), chunkBox, 48, 2, 1);
-        }
-
     }
 
     /**
@@ -150,12 +143,6 @@ public class GoldDungeonPieces {
 
         private static StructurePlaceSettings makeSettings() {
             return new StructurePlaceSettings().addProcessor(VerticalGradientProcessor.INSTANCE);
-        }
-
-        @Override
-        public void postProcess(WorldGenLevel level, StructureManager structureManager, ChunkGenerator generator, RandomSource random, BoundingBox chunkBox, ChunkPos pChunkPos, BlockPos pPos) {
-            super.postProcess(level, structureManager, generator, random, chunkBox, pChunkPos, pPos);
-            placeGoldenOaks(level, generator, random, this.getBoundingBox(), chunkBox, 64, 1, 0);
         }
     }
 
@@ -278,58 +265,5 @@ public class GoldDungeonPieces {
         protected void handleDataMarker(String pName, BlockPos pPos, ServerLevelAccessor pLevel, RandomSource pRandom, BoundingBox pBox) {
 
         }
-    }
-
-    /**
-     * Randomly place golden oak trees and flowers on top of a structure piece.
-     * @param boundingBox - The structure piece's bounding box
-     * @param chunkBox - The current chunk's bounding box
-     * @param randomBounds - The parameter for random.nextInt()
-     * @param treeWeight - The chance out of randomBounds of placing a tree
-     * @param flowerWeight - The chance out of randomBounds of placing a flower
-     */
-    private static void placeGoldenOaks(WorldGenLevel level, ChunkGenerator generator, RandomSource random, BoundingBox boundingBox, BoundingBox chunkBox, int randomBounds, int treeWeight, int flowerWeight) {
-        BlockPos.MutableBlockPos mutable = new BlockPos.MutableBlockPos();
-        int minX = Math.max(chunkBox.minX(), boundingBox.minX());
-        int minZ = Math.max(chunkBox.minZ(), boundingBox.minZ());
-        int maxX = Math.min(chunkBox.maxX(), boundingBox.maxX());
-        int maxZ = Math.min(chunkBox.maxZ(), boundingBox.maxZ());
-        int minY = boundingBox.minY() + Mth.floor((boundingBox.maxY() - boundingBox.minY()) * 0.75);
-        int maxY = boundingBox.maxY();
-
-        for (int x = minX; x < maxX; ++x) {
-            for (int z = minZ; z < maxZ; ++z) {
-                int featureType = random.nextInt(randomBounds);
-                if (featureType < treeWeight + flowerWeight) {
-                    mutable.set(x, maxY, z);
-                    if (iterateColumn(level, mutable, minY, maxY)) {
-                        if (featureType < treeWeight) {
-                            PlacedFeature tree = PlacementUtils.inlinePlaced(level.registryAccess().registryOrThrow(Registries.CONFIGURED_FEATURE).getHolderOrThrow(AetherConfiguredFeatures.GOLDEN_OAK_TREE_CONFIGURATION)).get();
-                            tree.place(level, generator, random, mutable);
-                        } else {
-                            Block flower = random.nextBoolean() ? Blocks.DANDELION : Blocks.POPPY;
-                            level.setBlock(mutable, flower.defaultBlockState(), 2);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Returns true if there is a solid block in the column. MutableBlockPos is set to the first empty block.
-     * @param level - The level to check for blocks.
-     * @param pos - This MutableBlockPos is set to the first empty block in the column.
-     */
-    private static boolean iterateColumn(WorldGenLevel level, BlockPos.MutableBlockPos pos, int minY, int maxY) {
-        int y;
-        for (y = maxY; y > minY; --y) {
-            pos.setY(y);
-            if (level.getBlockState(pos).is(AetherTags.Blocks.AETHER_DIRT)) {
-                pos.setY(++y);
-                return true;
-            }
-        }
-        return false;
     }
 }
